fixed score evaluator to show percentage properly

+++ src/main/java/org/example/score/ScoreEvaluator.java
@@ -2,9 +2,9 @@ package org.example.score;
 
 public class ScoreEvaluator {
 
-    public void evaluateScore(int score, int numberOfQuestions){
+    public void evaluateScore(double score, int numberOfQuestions){
         System.out.println("Konec kvízu.");
-        System.out.println("Skóre: " + score + "/" + numberOfQuestions);
+        System.out.println("Skóre: " + (int)score + "/" + numberOfQuestions);
         System.out.println("Úspěšnost: " + ((score / numberOfQuestions) * 100) + " %");
     }
 }
